move component tree view out into separate file
