+++ nexus_constructor/main_window.py
@@ -1,12 +1,4 @@
-from PySide2.QtCore import QModelIndex
-from PySide2.QtWidgets import (
-    QToolBar,
-    QAbstractItemView,
-    QMainWindow,
-    QApplication,
-    QInputDialog,
-    QLineEdit,
-)
+from PySide2.QtWidgets import QMainWindow, QApplication, QInputDialog, QLineEdit
 from PySide2.QtWidgets import QDialog, QLabel, QGridLayout, QComboBox, QPushButton
 
 import silx.gui.hdf5
@@ -17,19 +9,11 @@ import nexus_constructor.json.forwarder_json_writer
 from nexus_constructor.add_component_window import AddComponentDialog
 from nexus_constructor.filewriter_command_dialog import FilewriterCommandDialog
 from nexus_constructor.instrument import Instrument
-from nexus_constructor.main_window_utils import (
-    create_and_add_toolbar_action,
-    set_button_states,
-    expand_transformation_list,
-    add_transformation,
-)
-from nexus_constructor.transformation_types import TransformationType
 from nexus_constructor.ui_utils import file_dialog, show_warning_dialog
 from ui.main_window import Ui_MainWindow
 from nexus_constructor.component.component import Component
 
-from nexus_constructor.component_tree_model import ComponentTreeModel
-from nexus_constructor.component_tree_view import ComponentEditorDelegate
+
 from nexus_constructor.json import filewriter_json_writer
 from nexus_constructor.json.filewriter_json_reader import json_to_nexus
 
@@ -82,93 +66,12 @@ class MainWindow(Ui_MainWindow, QMainWindow):
         )
 
         self.widget.setVisible(True)
+        self.component_tree_view_tab.set_up_model(self.instrument)
 
-        self._set_up_tree_view()
         self.set_up_warning_window()
 
-    def _set_up_tree_view(self):
-        self._set_up_component_model()
-        self.component_tree_view.setDragEnabled(True)
-        self.component_tree_view.setAcceptDrops(True)
-        self.component_tree_view.setDropIndicatorShown(True)
-        self.component_tree_view.header().hide()
-        self.component_tree_view.updateEditorGeometries()
-        self.component_tree_view.updateGeometries()
-        self.component_tree_view.updateGeometry()
-        self.component_tree_view.clicked.connect(self._set_button_state)
-        self.component_tree_view.setSelectionMode(QAbstractItemView.SingleSelection)
-
-        self.component_tool_bar = QToolBar("Actions", self.component_tree_view_tab)
-        self.new_component_action = create_and_add_toolbar_action(
-            "new_component.png",
-            "New Component",
-            self.show_add_component_window,
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-            True,
-        )
-        self.new_translation_action = create_and_add_toolbar_action(
-            "new_translation.png",
-            "New Translation",
-            lambda: self._add_transformation(TransformationType.TRANSLATION),
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.new_rotation_action = create_and_add_toolbar_action(
-            "new_rotation.png",
-            "New Rotation",
-            lambda: self._add_transformation(TransformationType.ROTATION),
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.create_link_action = create_and_add_toolbar_action(
-            "create_link.png",
-            "Create Link",
-            self.on_create_link,
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.duplicate_action = create_and_add_toolbar_action(
-            "duplicate.png",
-            "Duplicate",
-            self.on_duplicate_node,
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.edit_component_action = create_and_add_toolbar_action(
-            "edit_component.png",
-            "Edit Component",
-            self.show_edit_component_dialog,
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.delete_action = create_and_add_toolbar_action(
-            "delete.png",
-            "Delete",
-            self.on_delete_item,
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.zoom_action = create_and_add_toolbar_action(
-            "zoom.svg",
-            "Zoom To Component",
-            self.on_zoom_item,
-            self.component_tool_bar,
-            self.component_tree_view_tab,
-        )
-        self.component_tool_bar.insertSeparator(self.zoom_action)
-        self.componentsTabLayout.insertWidget(0, self.component_tool_bar)
-
-    def _set_up_component_model(self):
-        self.component_model = ComponentTreeModel(self.instrument)
-        self.component_delegate = ComponentEditorDelegate(
-            self.component_tree_view, self.instrument
-        )
-        self.component_tree_view.setItemDelegate(self.component_delegate)
-        self.component_tree_view.setModel(self.component_model)
-
     def show_edit_component_dialog(self):
-        selected_component = self.component_tree_view.selectedIndexes()[
+        selected_component = self.component_tree_view_tab.component_tree_view.selectedIndexes()[
             0
         ].internalPointer()
         self.show_add_component_window(selected_component)
@@ -209,39 +112,6 @@ class MainWindow(Ui_MainWindow, QMainWindow):
         self.entries_dialog.layout().addWidget(combo)
         self.entries_dialog.layout().addWidget(ok_button)
         self.entries_dialog.show()
-
-    def _set_button_state(self):
-        set_button_states(
-            self.component_tree_view,
-            self.delete_action,
-            self.duplicate_action,
-            self.new_rotation_action,
-            self.new_translation_action,
-            self.create_link_action,
-            self.zoom_action,
-            self.edit_component_action,
-        )
-
-    def on_create_link(self):
-        selected = self.component_tree_view.selectedIndexes()
-        if len(selected) > 0:
-            self.component_model.add_link(selected[0])
-            self._expand_transformation_list(selected[0])
-            self._set_button_state()
-
-    def on_duplicate_node(self):
-        selected = self.component_tree_view.selectedIndexes()
-        if len(selected) > 0:
-            self.component_model.duplicate_node(selected[0])
-            self._expand_transformation_list(selected[0])
-
-    def _expand_transformation_list(self, node: QModelIndex):
-        expand_transformation_list(node, self.component_tree_view, self.component_model)
-
-    def _add_transformation(self, transformation_type: TransformationType):
-        add_transformation(
-            transformation_type, self.component_tree_view, self.component_model
-        )
 
     def set_up_warning_window(self):
         """
@@ -373,23 +243,10 @@ class MainWindow(Ui_MainWindow, QMainWindow):
         self.add_component_window = QDialog()
         self.add_component_window.ui = AddComponentDialog(
             self.instrument,
-            self.component_model,
+            self.component_tree_view_tab.component_model,
             component,
             definitions_dir=self.definitions_dir,
             parent=self,
         )
         self.add_component_window.ui.setupUi(self.add_component_window)
         self.add_component_window.show()
-
-    def on_delete_item(self):
-        selected = self.component_tree_view.selectedIndexes()
-        for item in selected:
-            self.component_model.remove_node(item)
-        self._set_button_state()
-
-    def on_zoom_item(self):
-        selected = self.component_tree_view.selectedIndexes()[0]
-        component = selected.internalPointer()
-        self.sceneWidget.zoom_to_component(
-            self.sceneWidget.get_entity(component.name), self.sceneWidget.view.camera()
-        )
